requirement.txt
imageio
Pillow
numpy
torch
torchvision
albumentations
tqdm
